migrate config sync module to a crate
