+++ src-tauri/core/src/connection/mod.rs
@@ -1,3 +1,4 @@
+pub mod active_connections;
 pub mod daemon_client;
 pub mod setup;
