+++ src-tauri/core/src/lib.rs
@@ -13,6 +13,7 @@ pub mod app_config;
 pub mod connection;
 pub mod database;
 pub mod error;
+pub mod events;
 pub mod proxy;
 pub mod version;
 pub mod wg_config;
