+++ src-tauri/enterprise/config-sync/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "defguard-config-sync"
+description = "Real-time configuration sync for the Defguard desktop client"
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license-file = "../LICENSE.md"
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+defguard-client-core = { path = "../../core" }
+defguard-client-proto = { path = "../../client-proto" }
+log = { version = "0.4", features = ["serde"] }
+reqwest = { version = "0.13", features = ["json"] }
+semver = { version = "1.0", features = ["serde"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+sqlx = { version = "0.8", features = ["chrono", "runtime-tokio", "sqlite"] }
+tauri = { version = "2", features = ["tray-icon"] }
+tokio = { version = "1", features = ["time"] }
+
+[target.'cfg(not(target_os = "macos"))'.dependencies]
+tonic = { workspace = true }
