+++ src-tauri/src/commands.rs
@@ -1,13 +1,8 @@
 use core::fmt;
-use std::{
-    collections::{HashMap, HashSet},
-    env,
-    str::FromStr,
-};
+use std::{collections::HashMap, env, str::FromStr};
 
 use chrono::{DateTime, Duration, Utc};
 use serde::{Deserialize, Serialize};
-use sqlx::{Sqlite, Transaction};
 use struct_patch::Patch;
 use tauri::{AppHandle, Emitter, Manager, State};
 
@@ -30,7 +25,9 @@ use crate::{
         DB_POOL,
     },
     enterprise::{
-        self, periodic::config::poll_instance, posture::authorize_posture_session,
+        self,
+        periodic::config::{do_update_instance, poll_instance},
+        posture::authorize_posture_session,
         provisioning::ProvisioningConfig,
     },
     error::Error,
@@ -673,239 +670,6 @@ pub async fn update_instance(
         error!("Instance to update with id {instance_id} was not found, aborting update");
         Err(Error::NotFound)
     }
-}
-
-/// Returns true if configuration in instance_info differs from current configuration
-pub(crate) async fn locations_changed(
-    transaction: &mut Transaction<'_, Sqlite>,
-    instance: &Instance<Id>,
-    device_config: &DeviceConfigResponse,
-) -> Result<bool, Error> {
-    let db_locations = Location::find_by_instance_id(transaction.as_mut(), instance.id, true)
-        .await?
-        .into_iter()
-        .map(|location| {
-            let mut new_location = Location::<NoId>::from(location);
-            // Ignore `route_all_traffic` flag as Defguard core does not have it.
-            new_location.route_all_traffic = false;
-            // Canonicalize mfa_method so a user-set value doesn't falsely trigger a
-            // config-change detection when the mode hasn't actually changed.
-            new_location.mfa_method = infer_mfa_method(new_location.location_mfa_mode, None);
-            new_location
-        })
-        .collect::<HashSet<_>>();
-    let core_locations: HashSet<Location> = device_config
-        .configs
-        .iter()
-        .map(|config| into_location(config.clone(), instance.id))
-        .collect::<HashSet<_>>();
-
-    Ok(db_locations != core_locations)
-}
-
-pub(crate) async fn do_update_instance(
-    transaction: &mut Transaction<'_, Sqlite>,
-    instance: &mut Instance<Id>,
-    response: DeviceConfigResponse,
-) -> Result<(), Error> {
-    // update instance
-    debug!("Updating instance {instance}");
-    let locations_changed = locations_changed(transaction, instance, &response).await?;
-    let instance_info = response
-        .instance
-        .expect("Missing instance info in device config response");
-    instance.name = instance_info.name;
-    instance.url = instance_info.url;
-    instance.proxy_url = instance_info.proxy_url;
-    instance.username = instance_info.username;
-    // Make sure to update the locations too if we are disabling all traffic
-    let policy = instance_info.client_traffic_policy.into();
-    if instance.client_traffic_policy != policy && policy == ClientTrafficPolicy::DisableAllTraffic
-    {
-        debug!("Disabling all traffic for all locations of instance {instance}");
-        Location::disable_all_traffic_for_all(transaction.as_mut(), instance.id).await?;
-        debug!("Disabled all traffic for all locations of instance {instance}");
-    }
-    instance.client_traffic_policy = instance_info.client_traffic_policy.into();
-    instance.openid_display_name = instance_info.openid_display_name;
-    instance.uuid = instance_info.id;
-    // Token may be empty if it was not issued
-    // This happens during polling, as core doesn't issue a new token for polling request
-    if response.token.is_some() {
-        instance.token = response.token;
-        debug!("Set polling token for instance {}", instance.name);
-    } else {
-        debug!(
-            "No polling token received for instance {}, not updating",
-            instance.name
-        );
-    }
-    instance.save(transaction.as_mut()).await?;
-    debug!(
-        "A new base configuration has been applied to instance {instance}, even if nothing changed"
-    );
-
-    let mut service_locations = Vec::new();
-
-    // check if locations have changed
-    if locations_changed {
-        // process locations received in response
-        debug!(
-            "Updating locations for instance {}({}).",
-            instance.name, instance.id
-        );
-        // Fetch existing locations for a given instance.
-        let mut current_locations =
-            Location::find_by_instance_id(transaction.as_mut(), instance.id, true).await?;
-        for dev_config in response.configs {
-            // parse device config
-            let new_location = into_location(dev_config, instance.id);
-
-            // check if location is already present in current locations
-            let saved_location = if let Some(position) = current_locations
-                .iter()
-                .position(|loc| loc.network_id == new_location.network_id)
-            {
-                // remove from list of existing locations
-                let mut current_location = current_locations.remove(position);
-                debug!(
-                    "Updating existing location {}({}) for instance {}({}).",
-                    current_location.name, current_location.id, instance.name, instance.id,
-                );
-                // update existing location
-                current_location.name = new_location.name;
-                current_location.address = new_location.address;
-                current_location.pubkey = new_location.pubkey;
-                current_location.endpoint = new_location.endpoint;
-                current_location.allowed_ips = new_location.allowed_ips;
-                current_location.keepalive_interval = new_location.keepalive_interval;
-                current_location.dns = new_location.dns;
-                current_location.location_mfa_mode = new_location.location_mfa_mode;
-                current_location.service_location_mode = new_location.service_location_mode;
-                // Correct mfa_method to remain consistent with the (possibly updated) mfa_mode.
-                current_location.mfa_method = infer_mfa_method(
-                    current_location.location_mfa_mode,
-                    current_location.mfa_method,
-                );
-                current_location.posture_check_required = new_location.posture_check_required;
-                current_location.save(transaction.as_mut()).await?;
-                info!("Location {current_location} configuration updated for instance {instance}");
-                current_location
-            } else {
-                // create new location
-                debug!("Creating new location {new_location} for instance instance {instance}");
-                let new_location = new_location.save(transaction.as_mut()).await?;
-                info!("New location {new_location} created for instance {instance}");
-                new_location
-            };
-
-            if saved_location.is_service_location() {
-                debug!(
-                    "Adding service location {}({}) for instance {}({}) to be saved to the daemon.",
-                    saved_location.name, saved_location.id, instance.name, instance.id,
-                );
-                service_locations.push(crate::enterprise::service_locations::to_service_location(
-                    &saved_location,
-                )?);
-            }
-        }
-
-        // remove locations which were present in current locations
-        // but no longer found in core response
-        debug!("Removing locations for instance {instance}");
-        for removed_location in current_locations {
-            removed_location.delete(transaction.as_mut()).await?;
-            info!(
-                "Removed location {removed_location} for instance {instance} during instance update"
-            );
-        }
-        debug!("Finished updating locations for instance {instance}");
-    } else {
-        info!("Locations for instance {instance} didn't change. Not updating them.");
-    }
-
-    if service_locations.is_empty() {
-        debug!(
-            "No service locations for instance {}({}), removing all existing service locations connections if there are any.",
-            instance.name, instance.id
-        );
-
-        #[cfg(not(target_os = "macos"))]
-        {
-            let delete_request = DeleteServiceLocationsRequest {
-                instance_id: instance.uuid.clone(),
-            };
-            DAEMON_CLIENT
-            .clone()
-            .delete_service_locations(delete_request)
-            .await
-            .map_err(|err| {
-                error!(
-                    "Error while deleting service locations from the daemon for instance {}({}): {err}",
-                    instance.name, instance.id,
-                );
-                Error::InternalError(err.to_string())
-            })?;
-            debug!(
-                "Successfully removed all service locations from daemon for instance {}({})",
-                instance.name, instance.id
-            );
-        }
-    } else {
-        debug!(
-            "Processing {} service location(s) for instance {}({})",
-            service_locations.len(),
-            instance.name,
-            instance.id
-        );
-
-        #[cfg(not(target_os = "macos"))]
-        {
-            let private_key = WireguardKeys::find_by_instance_id(transaction.as_mut(), instance.id)
-                .await?
-                .ok_or(Error::NotFound)?
-                .prvkey;
-
-            let save_request = SaveServiceLocationsRequest {
-                service_locations: service_locations.clone(),
-                instance_id: instance.uuid.clone(),
-                private_key,
-            };
-
-            debug!(
-                "Sending request to daemon to save {} service location(s) for instance {}({})",
-                save_request.service_locations.len(),
-                instance.name,
-                instance.id
-            );
-
-            DAEMON_CLIENT
-                .clone()
-                .save_service_locations(save_request)
-                .await
-                .map_err(|err| {
-                    error!(
-                    "Error while saving service locations to the daemon for instance {}({}): {err}",
-                    instance.name, instance.id,
-                );
-                    Error::InternalError(err.to_string())
-                })?;
-
-            info!(
-                "Successfully saved {} service location(s) to daemon for instance {}({})",
-                service_locations.len(),
-                instance.name,
-                instance.id
-            );
-
-            debug!(
-                "Completed processing all service locations for instance {}({})",
-                instance.name, instance.id
-            );
-        }
-    }
-
-    Ok(())
 }
 
 /// If `datetime` is Some, parses the date string, otherwise returns `DateTime` one hour ago.
